requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-genrativeai=0.3.2
google-genai==0.1.0
gTTS==2.4.0
requests==2.31.0
Pillow
Werkzeug==3.0.1
